chore: Integrations Refactor
connects to #48

Simplified code by removing indirection to Github and Slack
modules. This was mainly used for subbing in mock modules.
This change instead relies more on bypass to stub http requests
instead of using Mox. Tests are slightly more complicated in some
cases, but also more explicit. This should also allow for code
completion to work on the integration modules.